# backend/swagger.yaml
openapi: 3.0.0
info:
  title: PDF Signing Application API
  version: 1.0.0
  description: |
    A comprehensive API for digital PDF document signing and verification system with role-based access control.

    ## Authentication
    Most endpoints require JWT authentication. Use the `/api/auth/login` endpoint to get a token, then include it in the `Authorization` header as `Bearer <token>`.

    ## User Roles
    - **UPLOADER**: Can upload documents, verify signed documents, and reject documents
    - **SIGNER**: Can sign documents assigned to them

    ## Document Workflow
    1. Uploader uploads a PDF document and assigns it to a signer (status: PENDING)
    2. Signer signs the document with a digital signature (status: SIGNED)
    3. Uploader verifies or rejects the signed document (status: VERIFIED or REJECTED)
  contact:
    name: API Support
    email: support@pdfsigning.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.pdfsigning.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token in the format - Bearer <token>
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 0e370fcc-d2f6-4a95-82d3-a36a3ac5df0f
        name:
          type: string
          example: John Uploader
        email:
          type: string
          format: email
          example: uploader@example.com
        role:
          type: string
          enum: [UPLOADER, SIGNER]
          example: UPLOADER
        createdAt:
          type: string
          format: date-time
          example: 2025-10-13T23:00:00.000Z
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message here
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role
              properties:
                name:
                  type: string
                  minLength: 2
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                role:
                  type: string
                  enum: [uploader, signer, UPLOADER, SIGNER]
                  example: uploader
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      summary: Login user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: uploader@example.com
                password:
                  type: string
                  example: password123
                role:
                  type: string
                  enum: [uploader, signer]
                  example: uploader
                  description: Optional - validate that user has this role
            examples:
              uploader:
                summary: Login as uploader
                value:
                  email: uploader@example.com
                  password: password123
                  role: uploader
              signer:
                summary: Login as signer
                value:
                  email: signer@example.com
                  password: password123
                  role: signer
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials or wrong role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      summary: Get current user profile
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...